Added owner checking for solution templates

+++ src/main/java/com/sky1sbloo/ocjsys/code/problem/solutiontemplate/SolutionTemplateService.java
@@ -1,5 +1,6 @@
 package com.sky1sbloo.ocjsys.code.problem.solutiontemplate;
 
+import com.sky1sbloo.ocjsys.auth.AuthUser;
 import com.sky1sbloo.ocjsys.code.CodeLanguage;
 import com.sky1sbloo.ocjsys.code.problem.CodeProblem;
 import com.sky1sbloo.ocjsys.code.problem.CodeProblemRepository;
@@ -8,6 +9,7 @@ import com.sky1sbloo.ocjsys.code.problem.solutiontemplate.dto.SolutionTemplateEd
 import com.sky1sbloo.ocjsys.code.problem.solutiontemplate.dto.SolutionTemplateGetDto;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -36,16 +38,24 @@ public class SolutionTemplateService {
         return solutionTemplateRepository.findByCodeProblem_IdAndLanguage(problemId, language);
     }
 
-    public SolutionTemplate addSolutionTemplate(SolutionTemplateDto solutionTemplateCreateDto) {
+    public SolutionTemplate addSolutionTemplate(SolutionTemplateDto solutionTemplateCreateDto, AuthUser user)
+            throws AccessDeniedException {
         SolutionTemplate solutionTemplate = createSolutionTemplateFromDto(solutionTemplateCreateDto);
+        if (!userOwnsCodeProblem(user, solutionTemplate.getCodeProblem())) {
+            throw new AccessDeniedException("User do not own this code problem");
+        }
         return solutionTemplateRepository.save(solutionTemplate);
     }
 
-    public void editSolutionTemplate(Long id, SolutionTemplateEditDto solutionTemplateEditDto) {
+    public void editSolutionTemplate(Long id, SolutionTemplateEditDto solutionTemplateEditDto, AuthUser user)
+            throws AccessDeniedException {
         CodeProblem problem = codeProblemRepository.findById(id).orElseThrow(
                 () -> new EntityNotFoundException("Code problem with id: " + id +
                         " not found")
         );
+        if (!userOwnsCodeProblem(user, problem)) {
+            throw new AccessDeniedException("User do not own this code problem");
+        }
         SolutionTemplate solutionTemplate = SolutionTemplate.builder()
                 .codeProblem(problem)
                 .language(solutionTemplateEditDto.getLanguage())
@@ -55,9 +65,22 @@ public class SolutionTemplateService {
         solutionTemplateRepository.save(solutionTemplate);
     }
 
-    public void deleteSolutionTemplate(SolutionTemplateGetDto solutionTemplateGetDto) {
+    public void deleteSolutionTemplate(SolutionTemplateGetDto solutionTemplateGetDto, AuthUser user)
+            throws AccessDeniedException {
         SolutionTemplateId templateId = createTemplateIdFromDto(solutionTemplateGetDto);
+        CodeProblem problem = codeProblemRepository.findById(solutionTemplateGetDto.getProblemId()).orElseThrow(
+                () -> new EntityNotFoundException("Code problem with id: " + solutionTemplateGetDto.getProblemId() +
+                        " not found")
+        );
+        if (!userOwnsCodeProblem(user, problem)) {
+            throw new AccessDeniedException("User do not own this code problem");
+        }
         solutionTemplateRepository.deleteById(templateId);
+    }
+
+    private Boolean userOwnsCodeProblem(AuthUser user, CodeProblem codeProblem) {
+        // TODO: check if user is admin and allow ownership
+        return codeProblem.getOwner().getAuthUser().getUsername().equals(user.getUsername());
     }
 
     private SolutionTemplate createSolutionTemplateFromDto(SolutionTemplateDto solutionTemplateCreateDto)
